// src/app/interfaces/IPasteles.ts
// Generated by https://Quicktype.in

import ICategoria from "./ICategoria"

export interface IPastel {
    Pastelid: string
    nombre: string
    thumbnailUrl: string
    Description: string
    precio: number
    categoria: ICategoria;
}

export interface IColumn {
    key: string;
    label: string;
  }
  const ColumPastel:IColumn[] = [
      {
          key: "isbn",
          label: "ISBN",
        },
        {
          key: "title",
          label: "TITLE",
        },
        {
          key: "page",
          label: "PAGE",
        },
        {
          key: "precio",
          label: "PRECIO",
        },
        {
          key: "thumbnailUrl",
          label: "Imagen",
        },
  ];
  
  export default ColumPastel;
  